backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
google-genai>=1.14.0
PyMuPDF>=1.25.0
python-multipart>=0.0.9
pydantic>=2.9.0
sse-starlette>=2.1.0
starlette>=0.41.3
